requirements.txt
python-telegram-bot==20.7 
python-dotenv==1.2.2 
requests==2.31.0 
python-whois==0.9.4 
playwright==1.40.0 
aiohttp==3.9.1 
